feat(alerts): open alert history to the SQL caller and document it
Ticket 04: app.alert_events joins the tenant-scoped read surface. The
init file grants SELECT to sql_api_role, adds the default-deny row
policy, and documents the three-part tenancy template with the reason
each part exists. Provisioning covers new organizations through
SQL_API_TENANT_TABLES, which moves to its own module so the schema-probe
message and the MCP tool description derive the readable-table list
instead of hand-syncing copies. A migration backfills the per-
organization policies for organizations provisioned before the change;
it prints the exact statements from system.users. The projection MV
needs no policy of its own: row policies apply on read of app.logs, and
the argument is recorded where the next table copies the template.

Ticket 05: the everr-use-telemetry skill gains rules/alert-history.md,
in lockstep with the design doc's Reference: six query rules, the event
type table with an honest no-writer note, the full column reference, and
seven worked queries, each verified verbatim through the SQL API
profile. The hold-duration join and the state view stay out until their
writers exist. The doc's Reference gains the reserved inhibition freeze
columns the DDL already carries, and its stale error-column blocker now
records the sanitizer that ticket 02 shipped.
